Moment of truth?
